+++ libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: csakamot <csakamot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 13:40:58 by sakamoto          #+#    #+#             */
-/*   Updated: 2023/05/20 17:19:34 by csakamot         ###   ########.fr       */
+/*   Updated: 2023/05/29 12:03:48 by csakamot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	ft_putstr_fd(char *str, int fd);
 
 void	ft_putstr_fd(char *str, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!str)
